Refactor some more.
