+++ frontend/lib/event_list_component.dart
@@ -1,9 +1,8 @@
-import 'dart:async';
-
 import 'package:angular2/core.dart';
 import 'package:angular2/router.dart';
 import 'package:angular2_components/angular2_components.dart';
 import 'package:frontend/store/regatta_store.dart';
+import 'package:frontend/store/regatta_action.dart';
 
 import 'models/event.dart';
