+++ frontend/lib/store/action_type.dart
@@ -1,0 +1,8 @@
+/// The type of actions on the TodoApp.
+enum ActionType {
+  /// Indicates to create a new event by title
+  createNewEvent,
+
+  /// Indicates to add a loaded event to the store
+  addEvent
+}
