+++ frontend/lib/store/regatta_store.dart
@@ -1,56 +1,10 @@
 import 'dart:async';
+import 'action_type.dart';
+import 'regatta_action.dart';
 import 'package:angular2/core.dart';
 import 'package:greencat/greencat.dart';
 import 'package:frontend/models/event.dart';
 import 'package:frontend/services/event_service.dart';
-import 'package:frontend/app_component.dart';
-
-enum ActionType { addEvent, addNewEvent }
-
-/// Actions to be triggered to the app store.
-abstract class RegattaAction<T> extends Action<ActionType> {
-  /// The payload to include.
-  @override
-  final T payload;
-
-  /// Creates a new instance.
-  RegattaAction(this.payload);
-}
-
-/// Action to add an Event.
-class AddEventAction extends RegattaAction<Event> {
-  ///
-  AddEventAction(Event payload) : super(payload);
-
-  @override
-  ActionType get type => ActionType.addEvent;
-}
-
-/// Action to request the add of a new Event.
-class AddNewEventAction extends RegattaAction<String> implements AsyncAction<ActionType> {
-  final EventService _event_service;
-
-  ///
-  AddNewEventAction(String payload)
-      : _event_service = AppComponent.myinjector.get(EventService),
-        super(payload);
-
-  @override
-  ActionType get type => ActionType.addNewEvent;
-
-  @override
-  Future call(MiddlewareApi api) {
-    return _event_service.addEvent(payload).then((event) {
-      api.dispatch(addEvent(event));
-    });
-  }
-}
-
-/// Utility function to trigger the addEvent action.
-RegattaAction<Event> addEvent(Event event) => new AddEventAction(event);
-
-/// Utility function to trigger the addEvent action.
-RegattaAction<String> addNewEvent(String name) => new AddNewEventAction(name);
 
 /// State of the regattaApp.
 class RegattaState {
@@ -84,7 +38,7 @@ Reducer<RegattaState, RegattaAction<dynamic>> regattaApp =
       print('Reducer: addEvent');
       final events = (new List<Event>.from(currentState.events)..add(action.payload)).toList(growable: false);
       return currentState.copy(events: events);
-    case ActionType.addNewEvent:
+    case ActionType.createNewEvent:
       print('Reducer: addNewEvent');
       return currentState.copy();
     default:
